+++ src/main/java/edu/usfca/cs/mr/snowdepth/SnowDepthMapper.java
@@ -1,5 +1,6 @@
 package edu.usfca.cs.mr.snowdepth;
 
+import org.apache.hadoop.io.FloatWritable;
 import org.apache.hadoop.io.IntWritable;
 import org.apache.hadoop.io.LongWritable;
 import org.apache.hadoop.io.Text;
@@ -12,7 +13,7 @@ import java.util.StringTokenizer;
  * Mapper: Reads line by line, split them into words. Emit <word, 1> pairs.
  */
 public class SnowDepthMapper
-extends Mapper<LongWritable, Text, Text, IntWritable> {
+extends Mapper<LongWritable, Text, Text, FloatWritable> {
 
     @Override
     protected void map(LongWritable key, Text value, Context context)
@@ -30,7 +31,7 @@ extends Mapper<LongWritable, Text, Text, IntWritable> {
             }
             else if(index==50)
             {
-                IntWritable record= new IntWritable(Integer.parseInt(token));
+                FloatWritable record= new FloatWritable(Float.parseFloat(token));
                 context.write(new Text(geoHash), record);
             }
